backend/requirements.txt
flask
openai
youtube-transcript-api
python-dotenv






